Show only count for likes and comments initially and get full likes and comments on deliberate api request

+++ server/apps/post/views.py
@@ -6,14 +6,14 @@ from django.contrib.auth.models import User
 from apps.user.models import UserInfo
 from .models import Post, Like
 from apps.user.serializers import UserSerializer, UserInfoSerializer
-from .serializers import PostCreateSerializer, get_serialized_user_posts, get_serialized_posts, LikeCreateSerializer
+from .serializers import PostCreateSerializer, get_serialized_user_posts, get_serialized_posts, get_serialized_post_likes, LikeCreateSerializer
 
 from apps.user.utils import is_following
 from traceback import print_exc as print_error
 
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
-## Retrieve posts for test
+## Create post for user
 def post_create_view(request):
     try:
         requestingUser = request.user
@@ -66,22 +66,28 @@ def feed_view(request):
         return Response("Something went wrong", status=500)
     
 
-@api_view(['POST'])
+@api_view(['GET', 'POST'])
 @permission_classes([IsAuthenticated])
 def likes_view(request, pk):
     try:
         requestingUser = request.user
         targetPost = Post.objects.get(id=pk)
 
-        ##Since no request data so no need for a serializer
+        if request.method == 'POST':
+            ##Since no request data so no need for a serializer
 
-        liked = Like.objects.filter(liker=requestingUser, post=targetPost)
-        if liked:      #If this user already liked this post
-            liked.delete()
-            return Response("Like Deleted", status=200)
+            liked = Like.objects.filter(liker=requestingUser, post=targetPost)
+            if liked:      #If this user already liked this post
+                liked.delete()
+                return Response("Like Deleted", status=200)
+            else:
+                Like.objects.create(post=targetPost, liker=requestingUser)
+                return Response("Like created", status=201)
+        elif request.method == 'GET':
+            serialized_likes_data = get_serialized_post_likes(targetPost)
+            return Response(serialized_likes_data, status=200)
         else:
-            Like.objects.create(post=targetPost, liker=requestingUser)
-            return Response("Like created", status=201)
+            return Response("Invalid Request", status=400)
     except Exception as ex:
         print_error()
         return Response("Something went wrong", status=500)
